Commit_#2

+++ app/src/main/java/com/example/myapplication/ResponseDryRun.java
@@ -9,9 +9,9 @@ public class ResponseDryRun implements Serializable {
 	private String lattLong;
 
 	@SerializedName("woeid")
-	private int woeid;
+	private long woeid;
 
-	public ResponseDryRun(String lattLong, int woeid, String title, String locationType) {
+	public ResponseDryRun(String lattLong, long woeid, String title, String locationType) {
 		this.lattLong = lattLong;
 		this.woeid = woeid;
 		this.title = title;
@@ -28,7 +28,7 @@ public class ResponseDryRun implements Serializable {
 		return lattLong;
 	}
 
-	public int getWoeid(){
+	public long getWoeid(){
 		return woeid;
 	}
